Add files via upload
Keep both of these files in the same folder and run locally. To get the server running, you'll need to open the path in CMD. Then type "python server.py". This will start the server in a separate window, and you should see a message in the server window with the text "The server is running...". After the server is confirmed to be running, then you can run the client.py in your IDE (I use Pycharm). The data will print to the terminal in Pycharm with the client file. At the moment, I left the quiz like functionality so you can see how it works, and edit it anyway you would like. I also put comment lines showing where you can comment out if you just want to print the questions.